requirements.txt
matplotlib>=3.8.0
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.32.3
streamlit>=1.35.0
